Make history a core service, expected at /history
