+++ src/components/Graph/GraphDisplay.vue
