+++ src/components/Graph/plotly-factory.js
