+++ src/components/Graph/state.d.ts
@@ -1,8 +1,7 @@
-import { QueryParams, QueryTarget, DisplayNames } from '@/plugins/history/state';
+import { QueryParams, QueryTarget, DisplayNames } from '@/store/history/state';
 import { Layout } from 'plotly.js';
 
 export interface GraphConfig {
-  serviceId: string;
   layout: Partial<Layout>;
   params: QueryParams;
   targets: QueryTarget[];
