+++ src/plugins/history/Graph/state.d.ts
@@ -1,5 +1,5 @@
 import { DashboardItem } from '@/store/dashboards/state';
-import { GraphConfig } from '@/plugins/history/components/Graph/state';
+import { GraphConfig } from '@/components/Graph/state';
 
 export interface HistoryItem extends DashboardItem {
   config: GraphConfig;
