--- src/plugins/history/features/index.ts
@@ -1,8 +1,0 @@
-import { Feature } from '@/store/features/state';
-import Graph from './Graph';
-
-const features: { [id: string]: Feature } = {
-  Graph,
-};
-
-export default features;
