+++ src/store/history/getters.ts
@@ -1,9 +1,9 @@
-import { read } from '@/helpers/dynamic-store';
-import { Metric } from '@/plugins/history/state';
-import { HistoryState } from './state';
-import { RootStore } from '@/store/state';
+import { Metric, HistoryState } from './state';
+import { RootStore, RootState } from '@/store/state';
+import { getStoreAccessors } from 'vuex-typescript';
 
-export const typeName: string = 'History';
+const { read } = getStoreAccessors<HistoryState, RootState>('history');
+
 export const defaultMaxPoints: number = 500;
 
 const getters = {
@@ -22,24 +22,16 @@ export const metricValues = read(getters.metricValues);
 export const measurements = read(getters.measurements);
 export const fields = read(getters.fields);
 
-export const metricById = (store: RootStore, serviceId: string, metricId: string): Metric => {
-  const metric = metrics(store, serviceId)[metricId];
+export const metricById = (store: RootStore, metricId: string): Metric => {
+  const metric = metrics(store)[metricId];
   if (metric === undefined) {
-    throw new Error(`${metricId} not found in ${serviceId} store`);
+    throw new Error(`${metricId} not found in history store`);
   }
   return metric;
 };
 
-export const tryMetricById = (
-  store: RootStore,
-  serviceId: string,
-  metricId: string,
-): Metric | null =>
-  metrics(store, serviceId)[metricId] || null;
+export const tryMetricById = (store: RootStore, metricId: string): Metric | null =>
+  metrics(store)[metricId] || null;
 
-export const fieldsByMeasurement = (
-  store: RootStore,
-  serviceId: string,
-  measurement: string,
-) =>
-  fields(store, serviceId)[measurement];
+export const fieldsByMeasurement = (store: RootStore, measurement: string) =>
+  fields(store)[measurement];
